slp/infranstructure/parser/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
trafilatura==1.6.1
pypdf==3.17.4
python-magic-bin==0.4.14
pydantic==2.5.2
aiofiles==23.2.1
httpx==0.25.1
python-dotenv==1.0.0
python-multipart
fitz
bs4